DATAGRAPH-1387 - Move content of repository.core back to .support.
