+++ src/main/java/org/springframework/data/neo4j/core/DefaultNeo4jClient.java
@@ -42,7 +42,7 @@ import org.springframework.dao.support.PersistenceExceptionTranslator;
 import org.springframework.data.neo4j.core.convert.Neo4jConversions;
 import org.springframework.data.neo4j.core.transaction.Neo4jTransactionManager;
 import org.springframework.data.neo4j.core.transaction.Neo4jTransactionUtils;
-import org.springframework.data.neo4j.repository.core.Neo4jPersistenceExceptionTranslator;
+import org.springframework.data.neo4j.repository.support.Neo4jPersistenceExceptionTranslator;
 import org.springframework.lang.Nullable;
 import org.springframework.util.Assert;
