+++ src/main/java/org/springframework/data/neo4j/repository/config/EnableNeo4jRepositories.java
@@ -29,7 +29,7 @@ import org.springframework.core.annotation.AliasFor;
 import org.springframework.data.neo4j.core.Neo4jTemplate;
 import org.springframework.data.neo4j.core.mapping.Neo4jMappingContext;
 import org.springframework.data.neo4j.core.transaction.Neo4jTransactionManager;
-import org.springframework.data.neo4j.repository.core.Neo4jRepositoryFactoryBean;
+import org.springframework.data.neo4j.repository.support.Neo4jRepositoryFactoryBean;
 import org.springframework.data.repository.config.DefaultRepositoryBaseClass;
 
 /**
