--- src/main/java/org/springframework/data/neo4j/repository/core/package-info.java
@@ -1,7 +1,0 @@
-/**
- * This package contains infrastructure code that set up the needed Spring beans to register the repository capabilities.
- */
-@NonNullApi
-package org.springframework.data.neo4j.repository.core;
-
-import org.springframework.lang.NonNullApi;
