+++ src/main/java/org/springframework/data/neo4j/repository/support/Neo4jRepositoryFactorySupport.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.data.neo4j.repository.core;
+package org.springframework.data.neo4j.repository.support;
 
 /**
  * Shared repository factory functionality between imperative and reactive world.
