+++ src/main/java/org/springframework/data/neo4j/repository/support/ReactivePersistenceExceptionTranslationInterceptor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.data.neo4j.repository.core;
+package org.springframework.data.neo4j.repository.support;
 
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
