+++ src/main/java/org/springframework/data/neo4j/repository/support/SimpleNeo4jRepository.java
@@ -32,7 +32,6 @@ import org.springframework.data.neo4j.core.Neo4jOperations;
 import org.springframework.data.neo4j.core.mapping.Neo4jPersistentEntity;
 import org.springframework.data.neo4j.core.mapping.CypherGenerator;
 import org.springframework.data.neo4j.repository.query.CypherAdapterUtils;
-import org.springframework.data.neo4j.repository.core.Neo4jEntityInformation;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.support.PageableExecutionUtils;
 import org.springframework.stereotype.Repository;
