+++ src/main/java/org/springframework/data/neo4j/repository/support/SimpleReactiveNeo4jRepository.java
@@ -30,7 +30,6 @@ import org.springframework.data.neo4j.core.ReactiveNeo4jOperations;
 import org.springframework.data.neo4j.core.mapping.Neo4jPersistentEntity;
 import org.springframework.data.neo4j.core.mapping.CypherGenerator;
 import org.springframework.data.neo4j.repository.query.CypherAdapterUtils;
-import org.springframework.data.neo4j.repository.core.Neo4jEntityInformation;
 import org.springframework.data.repository.reactive.ReactiveSortingRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
