+++ src/test/java/org/springframework/data/neo4j/integration/imperative/CustomBaseRepositoryIT.java
@@ -34,7 +34,7 @@ import org.springframework.data.neo4j.core.Neo4jOperations;
 import org.springframework.data.neo4j.integration.shared.PersonWithAllConstructor;
 import org.springframework.data.neo4j.repository.Neo4jRepository;
 import org.springframework.data.neo4j.repository.config.EnableNeo4jRepositories;
-import org.springframework.data.neo4j.repository.core.Neo4jEntityInformation;
+import org.springframework.data.neo4j.repository.support.Neo4jEntityInformation;
 import org.springframework.data.neo4j.repository.support.SimpleNeo4jRepository;
 import org.springframework.data.neo4j.test.DriverMocks;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
