+++ src/test/java/org/springframework/data/neo4j/integration/reactive/ReactiveExceptionTranslationIT.java
@@ -41,8 +41,8 @@ import org.springframework.data.neo4j.core.schema.Id;
 import org.springframework.data.neo4j.core.schema.Node;
 import org.springframework.data.neo4j.repository.ReactiveNeo4jRepository;
 import org.springframework.data.neo4j.repository.config.EnableReactiveNeo4jRepositories;
-import org.springframework.data.neo4j.repository.core.Neo4jPersistenceExceptionTranslator;
-import org.springframework.data.neo4j.repository.core.ReactivePersistenceExceptionTranslationPostProcessor;
+import org.springframework.data.neo4j.repository.support.Neo4jPersistenceExceptionTranslator;
+import org.springframework.data.neo4j.repository.support.ReactivePersistenceExceptionTranslationPostProcessor;
 import org.springframework.data.neo4j.test.Neo4jExtension;
 import org.springframework.data.neo4j.test.Neo4jIntegrationTest;
 import org.springframework.stereotype.Repository;
