+++ src/test/java/org/springframework/data/neo4j/repository/support/Neo4jRepositoryFactorySupportTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.data.neo4j.repository.core;
+package org.springframework.data.neo4j.repository.support;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.assertj.core.api.Assertions.fail;
